Split classes in proper files
